refactor: styled components moved into _styled dir
